Separate the TCP and terse JSON formatting code.
